[BugFix] fix hour_from_unixtime rule (#63006)
(cherry picked from commit 3688bc4bd426a8fde0a3cf97c369e9c9aedd2220)

# Conflicts:
#	be/src/exprs/time_functions.cpp
#	fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/scalar/SimplifiedPredicateRule.java
#	fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rewrite/scalar/SimplifiedPredicateRuleTest.java

+++ gensrc/script/functions.py
@@ -565,7 +565,7 @@ vectorized_functions = [
     # TODO: 50380 year_from_unixtime
     # TODO: 50381 month_from_unixtime
     # TODO: 50382 day_from_unixtime
-    [50383, 'hour_from_unixtime', True, False, 'INT', ['BIGINT'], 'TimeFunctions::hour_from_unixtime'],
+    [50383, 'hour_from_unixtime', True, False, 'TINYINT', ['BIGINT'], 'TimeFunctions::hour_from_unixtime'],
     # TODO: 50384 minute_from_unixtime
     # TODO: 50385 second_from_unixtime
